+++ src/GitHub_Updater/API.php
@@ -510,8 +510,8 @@ class API {
 			return $response;
 		}
 
-		if ( is_dir( $repo->local_path ) &&
-			file_exists( $repo->local_path . $file )
+		if ( is_dir( $repo->local_path )
+			&& file_exists( $repo->local_path . $file )
 		) {
 			$response = file_get_contents( $repo->local_path . $file );
 		}
@@ -609,6 +609,7 @@ class API {
 	 * @uses  Requests, requires WP 4.6
 	 *
 	 * @param string $asset Release asset URI from git host.
+	 * @param bool   $aws   Release asset hosted on AWS.
 	 *
 	 * @return string|bool|\stdClass Release asset URI from AWS.
 	 */