+++ uninstall.php
@@ -37,7 +37,8 @@ global $wpdb;
 $table         = is_multisite() ? $wpdb->base_prefix . 'sitemeta' : $wpdb->base_prefix . 'options';
 $column        = is_multisite() ? 'meta_key' : 'option_name';
 $delete_string = 'DELETE FROM ' . $table . ' WHERE ' . $column . ' LIKE %s LIMIT 1000';
-// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
+// phpcs:disable
 $wpdb->query( $wpdb->prepare( $delete_string, [ '%ghu-%' ] ) );
+// phpcs:enable
 
 @unlink( WP_CONTENT_DIR . '/tmp-readme.txt' );