+++ CHANGES.md
@@ -5,6 +5,11 @@
 * update calls for Bitbucket Server REST API v7, thanks @Idealien
 * explicitly ignore themes without a root `style.css` file to avoid PHP warnings, thanks @cliffordp
 * move `Settings` action link to front
+* add Bitbucket pseudo-token, `username:password` for some private repos
+* Bitbucket credentials will automatically be converted to pseudo-tokens
+* update WP-CLI integration for Bitbucket pseudo-token
+* add plugins without updates to `$transient->no_update` to add _View details_ link, thanks @robincornett
+* no longer need to test if private repo when sending auth headers, auth headers are always sent
 
 #### 9.3.2 / 2020-02-19
 * fixed some PHP warnings and 401 errors when access tokens not set in `Basic_Auth_Loader`